requirements.txt
PyPDF2==3.0.1
pandas==2.0.3
tabula-py==2.7.0
openpyxl==3.1.2
PyQt5==5.15.9